api-macros: Split response code generation into more files
